Fix SSRF vulnerability with reqwest HTTP client.
By default, reqwest follows redirects, which potentially
allows the authorization server to mount an SSRF attack on the
client by returning a 3xx redirect to an arbitrary internal
URL. This change disables redirects, which matches the default
behavior when using the curl HTTP client.

Thanks to @d0nutptr for helping to discover this issue!