Avoid URL canonicalization
The `url` crate performs canonicalization such as appending
a trailing slash to URLs that have no path component. This affects
certain URL comparisons that are required to be plain string
comparisons (e.g., redirect URIs). This change duplicates the URL
NewType code from the `openidconnect` crate to maintain an internal
string representation and avoid the canonicalization.