Allow non-https token revocation endpoint
HTTPS is still enforced for the `client.revoke_token()` method in compliance with RFC 7009, but may now be bypassed by calling `client.revoke_token_with_unchecked_url()`.

When developing my appliation locally, I run a local copy of the oauth server without https. It was surprising to find the https checking on the revocation request, since it isn't enforced for auth or token exchange requests.