Use DeviceCodeErrorResponseType::ExpiredToken for expired Device token
It is good to use RequestTokenError::ServerResponse for expired
Device token instead of having a "Device code expired" string
is set to Other(). We see to it that they are have the same Error
with the error came from the endpoint. There was a timing
issue found that expired_token is returned from the endpoint
and sometimes the RequestTokenError::Other() was returned
after the expiration time given by the endpoint.

We make it uniform for easy error handling upon use.