Add cancellation when polling of access token
There is a use case that a user application in the
device code workflow that cancels during login phase

The polling of access token is immediately start running
after the user code and the url was already given to the
user.

The idea of this is that the user application can create
their own http_client and be able to return an Error
result when cancelling.