Fixing sleep interval that compounds
It was observe that during polling of token, if an error
keeps returning from the http_client, the sleep interval
compounds by 2. This scenario can be reproduced if ethernet cable is
unplugged for a long time not more than the expiry time.

To reproduced, set up a custom curl http_client with connect_timeout()
and timeout() from Easy2 curl-rust crate.

Thus, a max back off interval is introduce to put a limit to
the compounding of time interval.