reqwest: Migrate to shared `Error` type and use `thiserror`'s `From` impl
`reqwest` no longer has a separate error type for sync and async
implementations.  In addition `thiserror` can generate `From`
implementations for automatic conversions in the `Try` (`?`) operator.

Finally, delete some unnecessary `#[cfg]` wraps that are already
guarding the module as a whole, and that the code won't compile without
anyway.